Factor out Auth Check and add before + after hooks
The default auth check is same as before but it can now be overridden by the user. The method gets (fromUser, toUser) args - although toUser is not used by the default auth check it could be in specific apps. Also note: the auth check now runs *every time* impersonate is called, not only the first time like before, because the permissions might depend on toUser.